+++ NEWS.md
@@ -5,6 +5,7 @@
 # Version 3.1.0
 - Herbaceous information in summary.forest
 - Herbaceous transpiration added to the water balance
+- Weather input can now have dates as Date or POSIXct objects in column 'dates' instead of row names
 
 # Version 3.0.0
 - Leaf area limited by allometries